+++ src/src/components/common/form/LabeledInput.jsx
@@ -4,8 +4,9 @@ import styled from "styled-components";
 
 const Styled = {
   Container: styled.div`
-    width: 22rem;
-    margin: 1rem 0;
+    width: 100%;
+    max-width: 22rem;
+    margin: 0.5rem 0;
   `,
   Input: styled.input`
     width: 100%;
@@ -18,6 +19,10 @@ const Styled = {
     &::placeholder {
       color: ${({ theme }) => theme.color.inactive};
       font-size: 1rem;
+
+      white-space: nowrap;
+      overflow: hidden;
+      text-overflow: ellipsis;
     }
   `,
 
@@ -41,9 +46,13 @@ const Styled = {
 /**
  * LabeledInput - 라벨이 포함된 입력 폼 컴포넌트.
  * @param {object} props 전달되는 props
- * @param {object} props.validation 입력 유효성 검사 (React Hook Form의 register 함수와 함께 사용)
- * @param {string} props.errorMsg 입력 유효성 검사 오류 메시지 (선택 사항)
- * @param {string} props.requireMsg 필수 입력 메시지 (선택 사항)
+ * @param {string || number} id 인풋 아이디
+ * @param {string || number} label 라벨
+ * @param {string} type input의 type
+ * @param {string} placeholder placeholder
+ * @param {string} errorMsg 입력 유효성 검사 오류 메시지 (선택 사항)
+ * @param {string} requireMsg 필수 입력 메시지 (선택 사항)
+ * @param {object} validation 입력 유효성 검사 (React Hook Form의 register 함수와 함께 사용)
  * @returns {JSX.Element} LabeledInput 컴포넌트의 JSX 요소
  */
